+++ CHANGELOG.md
@@ -10,7 +10,7 @@ MolSimToolkit.jl Changelog
 [badge-fix]: https://img.shields.io/badge/Fix-purple.svg
 [badge-info]: https://img.shields.io/badge/Info-gray.svg
 
-Version 2.0.0-DEV
+Version 2.0.0
 --------------
 - ![FEATURE][badge-feature] `goto_frame!(sim, iframe)` function goes to the desired frame and returns the `Simulation` object.
 - ![BREAKING][badge-breaking] `atoms` function was removed in favor of `get_atoms`. Use `get_atoms(goto_frame!(sim, 5))` for example to retrive the vector of atoms with positions of frame 5.